assignment2/requirements.txt
haversine==2.8.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
numpy==1.26.0
pandas==2.1.1
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
